Return value type 'int' of method recv() changed to 'size_t'.

+++ socket.h
@@ -28,7 +28,7 @@ public:
     void bind(const sockaddr_in &addr, size_t addrlen);
     void listen(int n);
     size_t send(const void *buf, size_t len, int n);
-    int recv(void *buf, size_t len, int n);
+    size_t recv(void *buf, size_t len, int n);
     int connect(const sockaddr_in &addr, size_t addrlen);
     Socket accept(sockaddr_in &addr, socklen_t &addrlen);
